Do not adapt function reference if all arguments are simple positional
Otherwise it led to inconsistency in handling vararg parameters mapped
to arrays. For example, in the case

    fun target(vararg xs: Int) {}
    fun use(fn: (IntArray) -> Unit): Any = fn
    use(::target)

An instance of `CallableReferenceAdaptation` was recorded, even though
there's no adaptation happening: vararg mapped to array is a standard
behavior for references since 1.0. This could lead to subtle behavior
differences if `use` would try to cast the function object to
`KFunction`, which we're going to prohibit for adapted function
references in 1.4.